128.1. Moved logic from UpdatePersonalInfo to UpdatePersonalInfoContainer. Renamed class prefix in select from general-settings to settings
