+++ src/App.scss
@@ -24,7 +24,7 @@
 
 // === react-select ===
 
-.generalInfo_select {
+.settings_select {
   // the whole react-select component
   &__control {
     font-size: 18px;
