+++ src/components/Settings/GeneralSettings/UpdateMyPersonalInfoForm/UpdateMyPersonalInfoForm.tsx
@@ -1,86 +1,36 @@
 import s from '../../../../styles/formStyle.module.scss';
-import * as yup from 'yup';
 import { Form, Formik } from 'formik';
 import FormControl from '../../../common/FormControl/FormControl';
 import FormError from '../../../common/FormError/FormError';
 import FormStatus from '../../../common/FormStatus/FormStatus';
 import SubmitLoadingBtn from '../../../common/SubmitLoadingBtn/SubmitLoadingBtn';
-import { useAppDispatch, useAppSelector } from '../../../../hooks/hooks';
-import { PersonalInfoInitialValues } from '../../../../types/formikTypes';
-import { updateMyPersonalInfo } from '../../../../redux/myselfSlice';
-import { useState } from 'react';
-import { AttitudeTowardsEnum } from '../../../../types/reduxTypes/authSliceTypes';
+import {
+  AttitudeTowardsEnum,
+  User,
+} from '../../../../types/reduxTypes/authSliceTypes';
 import FormSelectControl from '../../../common/FormSelectControl/FormSelectControl';
-import { PersonalInfoFieldsToSend } from '../../../../types/reduxTypes/myselfSliceTypes';
+import { PersonalInfoInitialValues } from '../../../../types/formikTypes';
+import { PersonalInfoValidationShema } from './UpdateMyPersonalInfoFormContainer';
 
-const validationSchema = yup.object({});
+interface Props {
+  initialValues: PersonalInfoInitialValues;
+  validationSchema: PersonalInfoValidationShema;
+  selectOptions: any[];
+  currentUser: User;
+  isMyPersonalInfoFetching: boolean;
+  isUserPersonalInfoSuccessfulySent: boolean;
+  updateMyPersonalInfoErrorMsg: string;
+  onSubmit: (data: PersonalInfoInitialValues) => void;
+  onSmokingSelectChange: (newValue: any) => void;
+  onDrinkingSelectChange: (newValue: any) => void;
+}
 
-const UpdateMyPersonalInfoForm: React.FC = () => {
-  const currentUser = useAppSelector((state) => state.auth.user)!;
-  const { isMyPersonalInfoFetching, updateMyPersonalInfoErrorMsg } =
-    useAppSelector((state) => state.myslef);
-  const { isUserPersonalInfoSuccessfulySent } = useAppSelector(
-    (state) => state.forms
-  );
-  const [selectSmokingAttidute, setSelectSmokingAttidute] = useState(
-    currentUser.personalInfo?.attitudeTowardsSmoking ||
-      AttitudeTowardsEnum.NOT_SELECTED
-  );
-  const [selectDrinkingAttitude, setSelectDrinkingAttidute] = useState(
-    currentUser.personalInfo?.attitudeTowardsDrinking ||
-      AttitudeTowardsEnum.NOT_SELECTED
-  );
-  
-  const selectOptions = [
-    { label: 'Negative', value: AttitudeTowardsEnum.NEGATIVE },
-    { label: 'Neutral', value: AttitudeTowardsEnum.NEUTRAL },
-    { label: 'Positive', value: AttitudeTowardsEnum.POSITIVE },
-    { label: 'Compromise', value: AttitudeTowardsEnum.COMPROMISE },
-    { label: 'Not selected', value: AttitudeTowardsEnum.NOT_SELECTED },
-  ];
-
-  const dispatch = useAppDispatch();
-
-  const initialValues: PersonalInfoInitialValues = {
-    aboutMe: currentUser.personalInfo?.aboutMe || '',
-    activities: currentUser.personalInfo?.activities || '',
-    favoriteMovies: currentUser.personalInfo?.favoriteMovies || '',
-    favoriteMusic: currentUser.personalInfo?.favoriteMusic || '',
-    favouriteBooks: currentUser.personalInfo?.favouriteBooks || '',
-    interests: currentUser.personalInfo?.interests || '',
-  };
-
-  const onSmokingSelectChange = (newValue: any) => {
-    if (!newValue) {
-      return setSelectSmokingAttidute(AttitudeTowardsEnum.NOT_SELECTED);
-    }
-    setSelectSmokingAttidute(newValue.value);
-  };
-
-  const onDrinkingSelectChange = (newValue: any) => {
-    if (!newValue) {
-      return setSelectDrinkingAttidute(AttitudeTowardsEnum.NOT_SELECTED);
-    }
-    setSelectDrinkingAttidute(newValue.value);
-  };
-
-  const onSubmit = (data: PersonalInfoInitialValues) => {
-    const fieldsToSend: PersonalInfoFieldsToSend = {
-      personalInfo: {
-        ...data,
-        attitudeTowardsDrinking: selectDrinkingAttitude,
-        attitudeTowardsSmoking: selectSmokingAttidute,
-      },
-    };
-
-    dispatch(updateMyPersonalInfo(fieldsToSend));
-  };
-
+const UpdateMyPersonalInfoForm: React.FC<Props> = (props) => {
   return (
     <Formik
-      initialValues={initialValues}
-      onSubmit={onSubmit}
-      validationSchema={validationSchema}
+      initialValues={props.initialValues}
+      onSubmit={props.onSubmit}
+      validationSchema={props.validationSchema}
     >
       <Form className={s.form}>
         <FormControl
@@ -102,25 +52,25 @@ const UpdateMyPersonalInfoForm: React.FC = () => {
           labelClass={s.form__label}
         />
         <FormSelectControl
-          options={selectOptions}
-          classNamePrefix="generalInfo_select"
+          options={props.selectOptions}
+          classNamePrefix="settings_select"
           defaultValue={
-            currentUser.personalInfo?.attitudeTowardsSmoking ||
+            props.currentUser.personalInfo?.attitudeTowardsSmoking ||
             AttitudeTowardsEnum.NOT_SELECTED
           }
           labelText="Attidute towards smoking"
-          onChange={onSmokingSelectChange}
+          onChange={props.onSmokingSelectChange}
           placeholder="Select your attitude..."
         />
         <FormSelectControl
-          options={selectOptions}
-          classNamePrefix="generalInfo_select"
+          options={props.selectOptions}
+          classNamePrefix="settings_select"
           defaultValue={
-            currentUser.personalInfo?.attitudeTowardsDrinking ||
+            props.currentUser.personalInfo?.attitudeTowardsDrinking ||
             AttitudeTowardsEnum.NOT_SELECTED
           }
           labelText="Attidute towards drinking"
-          onChange={onDrinkingSelectChange}
+          onChange={props.onDrinkingSelectChange}
           placeholder="Select your drinking..."
         />
         <FormControl
@@ -174,18 +124,18 @@ const UpdateMyPersonalInfoForm: React.FC = () => {
             btnType="submit"
             btnText="Update contact information"
             btnFetchingText="Updating contact information"
-            isFetching={isMyPersonalInfoFetching}
+            isFetching={props.isMyPersonalInfoFetching}
             onSubmit={() => {}}
           />
           <FormStatus
-            isSuccessfulySent={isUserPersonalInfoSuccessfulySent}
+            isSuccessfulySent={props.isUserPersonalInfoSuccessfulySent}
             preloaderClass={s.form__preloader}
             msgClass={s.form__success}
           />
         </div>
         <FormError
           customClass={s.form__error}
-          errorMsg={updateMyPersonalInfoErrorMsg}
+          errorMsg={props.updateMyPersonalInfoErrorMsg}
         />
       </Form>
     </Formik>
